+++ project/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './components/app/app';
+import { App } from './components/App/App';
 import { nanoid } from '@reduxjs/toolkit';
 
 const movies = [...Array(20)].map(() => ({
